Moved request tests into separate files, added tests for all provided values.
